[RFR] Fix selecting language on analysis wizard, targets step
Follow-up to #1136

To make sure the test does not unselected the desired target
language that is automatically selected from the application's
tags (language-discovery automatically adds tags when the
application is created), switch from a `click` action to a
`check` action.  A bit more code is used to get to the input
checkbox so the `check()` will work.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>