Update a test's title so cypress-tags doesn't include it by accident
Since `cypress-tags` works as a preprocessor, all of the
test blocks need to have static tag and title arguments. One
test suit has a dynamic title (just a simple ternary is technically
dynamic) and was always being included.

Switching the `it()` title to be a template string allows the
preprocessor to see it as a string and properly apply the tags
from the `define()` block.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>