Fix login test by checking the UI's window._env configuration
Before all tests, request the app's index.html and look for
the `_env` string to parse.  It is stored as a base-64 encoded
JSON object.  The value holds the UI's configuration settings,
and most importantly if authorization is enabled.  This _env
object can be checked in tests and be able to do different
things based on how the UI being tested is configured.

Changes:
  - Adjust the login utils function to be auth aware.  When auth
    is turned on, handle an initial password that needs to be changed
    properly.

  - A `before()` block to find, parse and store the `_env` block

  - Since the login test works in AUTH=="false" and AUTH=="true"
    environments, move the login test to `@ci` so it always runs

  - Adjust the default e2e logging level to INFO so `cy.log()` calls
    show up by default in the cypress runner UI

  - Added `"baseUrl": "."` to tsconfig.json to resolve a warning message

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>