Update cypress libraries, use the esbuild runtime (#1550)
* Update cypress libraries, use the esbuild runtime

The primary goal is to adopt the esbuild runtime to speed
up the startup time of scripts.  Taking the startup time from
5 seconds down to 3 seconds for each spec file for >100 files
could reduce the the total runtime down by ~3 minutes.

The secondary goal is to review and update the cypress
libraries to make sure they're loaded properly and consistently.
Dropped `cypress-xpath` since it deprecated and isn't used
in any tests.

Finally, setup husky and lint-staged so prettier runs on
git commit to help reduce fails of the prettier.yaml workflow.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>

* fixup the prettier workflow, add engines to package.json

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>

---------

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>