[RFR] Fix login test by checking the UI's window._env configuration (#1543)
* Fix login test by checking the UI's window._env configuration

Before all tests, request the app's index.html and look for
the `_env` string to parse.  It is stored as a base-64 encoded
JSON object.  The value holds the UI's configuration settings,
and most importantly if authorization is enabled.  This _env
object can be checked in tests and be able to do different
things based on how the UI being tested is configured.

Changes:
  - Adjust the login utils function to be auth aware.  When auth
    is turned on, handle an initial password that needs to be changed
    properly.

  - A `before()` block to find, parse and store the `_env` block

  - Since the login test works in AUTH=="false" and AUTH=="true"
    environments, move the login test to `@ci` so it always runs

  - Adjust the default e2e logging level to INFO so `cy.log()` calls
    show up by default in the cypress runner UI

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>

* Move the _env lookup to beforeEach() since aliases are reset before each test

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>

* Create a custom command uiEnvironmentConfig

The custom command will run and return the _env
object in a Chain so it can be used the same way
as any other cypress value.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>

* Move the commands and types to commands.ts

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>

---------

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>